requirements.txt
Flask==3.1.0
asyncua==1.1.5
paho-mqtt==2.1.0
psutil==6.1.0
docker==7.1.0
